requirements.txt
langchain
langchain_community
bs4
pypdf
python-dotenv
ipykernel
pandas
ipywidgets
chromadb
sentence-transformers
uvicorn
fastapi
ipykernel
python-box
pymilvus
redis
langchain-google-genai
langchain-experimental
langchain-core
pydantic==1.10
streamlit
tools
typing
langchain_openai
pydantic
faiss-cpu
redis
yfinance
duckduckgo-search
langchainhub
datasets==2.16.0
pyarrow
pinecone-client
transformers
langchain_anthropic
numexpr
langcorn
